chore: separated Vector2 and WorldToCanvas into different folders
because they have different dependencies, all so that the import statement can be shorter in Main.tsx
