+++ src/core/architecture/index.ts
@@ -1,5 +1,6 @@
-export * from "./components"
 export * from "./Time"
-export * from "./GameEngine"
+export * from "./components"
 export * from "./GameObject"
 export * from "./Scene"
+export * from "./InputSystem"
+export * from "./GameEngine"
