+++ src/core/helpers/WorldToCanvas.ts
@@ -1,5 +1,5 @@
 import { GameEngine } from "../architecture";
-import { Vector2 } from "./Vector2";
+import { Vector2 } from "../types";
 
 export function WorldToCanvas(worldPosition: Vector2, canvasSize: Vector2): Vector2 {
     // currently only transforms y position
