+++ src/core/index.ts
@@ -1,3 +1,4 @@
+export * from "./types"
 export * from "./architecture"
 export * from "./hooks"
-export * from "./utilities"
+export * from "./helpers"
