--- src/core/utilities/index.ts
@@ -1,2 +1,0 @@
-export * from "./Vector2"
-export * from "./Utilities"
